chore: move to `src/`
